main.star: triple sharding amount for misc-compile builders
It completes in under 30 minutes as is. Go with a tentative target
of 10 minutes for now (via golang/go#17104), we can revisit it later.

In comparison, the previous build system currently uses effectively
the equivalent of 33 shards (that each handle one port), so even 12
shards is only about a third of that.

For golang/go#61698.
For golang/go#17104.

Change-Id: Ie641e0574127fcb72b081219f540fb5e7a9cf020
Reviewed-on: https://go-review.googlesource.com/c/build/+/515735
TryBot-Bypass: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Michael Knyszek <mknyszek@google.com>