cmd/golangbuild: create directory linking to its location
The golangbuild command doesn't live in x/build at this time, but
it's expected to be eventually moved here. Point to its location
in the meantime.

This directory exists mostly to reserve the import path, which is
useful when filing issues in the public Go issue tracker pointing
to it.

Change-Id: I8cd3cb5d6cc65005f44a8c591dc19327a0493bc2
Reviewed-on: https://go-review.googlesource.com/c/build/+/518795
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Bypass: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>