gerrit: delete GetProjects in favor of ListProjects
The ListProjects and GetProjects methods were added in CL 38717 and
CL 37252 at around the same time, and they're both exposing the same
GET /projects/ Gerrit API endpoint. The latter is hardcoded to provide
the 'b' query parameter, one of many that GET /projects/ offers.

Rewrite cmd/gitmirror and cmd/updatestd to use ListProjects and N calls
to GetBranch instead, and delete GetProjects since we no longer use it.

I considered deprecating GetProjects instead of deleting, but x/build
and x/build/gerrit specifically are both documented to be for internal
Go project needs and not subject to the Go 1 compatibility promise, so
skip the overhead of doing that. If it happens to cause a problem that
can't be worked around in a better way, we'll consider re-adding it in
deprecated form. Add a pkg.go.dev link to increase visibility to the
general-purpose Gerrit API clients written in Go.

Updates golang/go#18743.

Change-Id: I583356526af04e915bce5464004eb68a650221d5
Reviewed-on: https://go-review.googlesource.com/c/build/+/518676
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>