cmd/gerritbot: fix ID used to post comments with PR rule findings
Use the more modern "<project>~<changeNumber>" format for a change
and properly specify the first revision.

While we are here, we add a few comments to the gerrit package
to help avoid confusion in the future.

This is a follow-up to CL 513397, which was failing to post the
new PR rule findings for a test PR with error:

2023/08/15 22:06:38 processPullRequest: importGerritChangeFromPR(golang/build#71, nil): could not add findings comment to CL for golang/build#71: HTTP status 404 Not Found on request to https://go-review.googlesource.com/a/changes/If1a8ae9e4b76a05b13139ddf9fda1cdf67b50b33/revisions/build~master~If1a8ae9e4b76a05b13139ddf9fda1cdf67b50b33/review; Not found: build~master~If1a8ae9e4b76a05b13139ddf9fda1cdf67b50b33

Updates golang/go#61573

Change-Id: Ib0fef88ae05b10e623c2c1af614aa932dbfcc043
Reviewed-on: https://go-review.googlesource.com/c/build/+/519796
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: t hepudds <thepudds1460@gmail.com>