internal/relui: add a solo Go 1.19 next minor workflow
In general, try to reuse as much existing code as possible to do this
one off-cycle release. Pretending that it's a KindPrevMinor release
kind (rather than creating a new KindPrevPrevMinor) should do that.

My plan is to use it up to the point of "Wait for Coordinator Approval"
to pre-build, and on the target release day, up to "Wait to Announce".

At that point, we can either let it send the rare but supported
"solo minor release" communication, or, perhaps better, instead
rely on the dual workflow for the Go 1.21.1 and Go 1.20.8 minor
releases to take care of the announcement. That is, we can add
a brief sentence to the announcement email and likely not worry
too much about the tweet.

For golang/go#62076.

Change-Id: Ie0783da9d4451a91311d99ccccf2689268ef39cf
Reviewed-on: https://go-review.googlesource.com/c/build/+/520236
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Heschi Kreinick <heschi@google.com>