main.star: refactor out an environment type
As I added the security trybots, the signature of define_builder got
bigger and bigger, and I promised myself that if I needed another
argument I'd refactor out a bag of properties. That time has come.

LOW_CAPACITY_HOSTS is a little strange because it applies to both
public try and ci builders, but I still put it in env just to avoid
having to pass it down. For now I'm using the same list for security to
prevent diffs, and because we don't have any Macs; once we do we can
change the rules just for the internal builds.

Change-Id: I4346eb9902df99ad42eeb08df4993e1f234c7771
Reviewed-on: https://go-review.googlesource.com/c/build/+/520877
Auto-Submit: Heschi Kreinick <heschi@google.com>
TryBot-Bypass: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>