internal/releasetargets: set GOARM=6 for linux-armv6l
From Dmitri's well-researched comment on golang/go#62164:

It was set to GOARM=6 in 1.20 and older because the release target used
the linux-arm-aws builder, and that builder has GOARM=6 in its
environment. For Go 1.21.0, it's built on a linux-amd64 builder via
make.bash -distpack, and GOARM wasn't set in the release target's
environment explicitly, so it ended up defaulting to 7 in make.bash.

Since this was always the effective environment, set it on the target
going back to 1.19. (I didn't get rid of 1.19 because we're still
planning the out-of-cycle release.)

Fixes golang/go#62164

Change-Id: I539d9dfaf9036e4ac9a6a11fc1e73440d8437a80
Reviewed-on: https://go-review.googlesource.com/c/build/+/521555
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Run-TryBot: Heschi Kreinick <heschi@google.com>
Auto-Submit: Heschi Kreinick <heschi@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>