main.star: enable x trybot coverage for supported Go versions, take 2
This is like CL 521935, which was rolled back in CL 522238, but this
time it takes advantage of knowing that CQ groups cannot overlap in
watched refs.

So it adds existing -go1-{n} builders to existing "x-repo-gotip" CQ
groups via additional cq_tryjob_verifier calls.

As before, this trybot coverage is limited to fast and available
linux/amd64 builders only.

If this works, it'll be possible to refactor and clean up our loop
in _define_go_ci without changing the generated configuration. This
CL aims to be minimal to find out if this approach will work.

Change-Id: I6850506d76e1181efdc74144a48c622692f9ecd7
Reviewed-on: https://go-review.googlesource.com/c/build/+/522515
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
TryBot-Bypass: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>