main.star: disable reuse of trybot results for all builders
Currently CV will cache results even in the event of a rebase. This is
may be an OK configuration if we start using CQ+2, though it might still
not be desirable when rebasing stacks of dependent changes.

In any case, for now, we should definitely disable this caching.

Change-Id: Ida6128bd1485fc8a98d9b2d3a15cdf35386d6775
Reviewed-on: https://go-review.googlesource.com/c/build/+/526735
Reviewed-by: Heschi Kreinick <heschi@google.com>
TryBot-Bypass: Michael Knyszek <mknyszek@google.com>