maintner/maintnerd/maintapi: test x/tools with Go 1.19+1.18 longer
We've been able to extend LUCI test coverage for the needs of gopls,
which makes Kokoro CI testing that previously provided said testing
obsolete. What I failed to realize until trying to clean up after it
is that it's still load bearing during the gradual transition to LUCI.
That is, someone may test a CL using the old trybots, get a passing
result, and submit.

That'll be quite annoying as it'd cause a breakage only when using the
new LUCI trybots, so I'd rather not leave this possibility in for long.

I can bring back the Kokoro CI and delete it later, or make x/tools
require LUCI trybots sooner, but perhaps a simpler path for now is to
manually extend coordinator trybots to include 1.19+1.18 in x/tools.
The same approach worked for x/website needs in CL 393858 and CL 456715.

Change-Id: If4f72049200ceb6e8d2286849d11cfdba98e4de2
Reviewed-on: https://go-review.googlesource.com/c/build/+/528155
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Heschi Kreinick <heschi@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>