internal/task: add script support to Cloud Build
We're seeing a proliferation of tasks that want to run a command or two
on an arbitrary worker machine. Using buildlets for that presents two
problems: it's an extremely awkward API, and we're planning to get rid
of them not so long from now.

Add support for running an arbitrary script to our Cloud Build wrapper
and switch all such tasks over to it. Next step will be the distpack
builds.

The RunScript API may be a bit overspecialized to the current use cases,
but it's nice to be able to replace a bunch of ceremony with a single
call. If we need something more flexible in the future we can do it
then.

For golang/go#63147

Change-Id: Iefcfae942c75816b8a49b1bd8ca06377f7246267
Reviewed-on: https://go-review.googlesource.com/c/build/+/529296
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>