main.star: permit sharding golang.org/x repos
The misc-compile builder is the bottleneck by a significant margin in
some golang.org/x repos such as x/tools. This is the first time we're
extending sharding beyond the main Go repo, so there are some changes
to builder definitions to support that. No changes to generated files
in this CL to prove it didn't break anything, and the next one starts
to adjust configuration to turn on sharding.

Note that sharding golang.org/x repo test execution requires that the
luciexe binary also supports it, which is the case for misc-compile.

Change-Id: Ic00dd7df8e85b7a3b611385664be335515cb3b1a
Reviewed-on: https://go-review.googlesource.com/c/build/+/528560
TryBot-Bypass: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>