internal/relui: use Cloud Build to build primary distpacks
Now that we have Cloud Build support, we can also use it in the Go
release process.

A few notes:
- We now use the latest Go release to bootstrap. I think this is fine?
- We still pass a tar of all the distpacks around. It might be possible
  to improve that, since GCB is capable of uploading multiple files
  easily enough, but I didn't want to make too many changes all at once.
- GCB needs to be able to read the relui scratch space, and write to its
  own.

We still need Windows buildlets to build MSIs and do the
reproducibility check, and Macs to build and extract packages. It might
be possible to do the packaging steps on Linux, but the reproducibility
check is pointless unless it really runs on something different. I
suspect we'll need to use the Swarming API for lack of other options.

For golang/go#63147

Change-Id: Ib4e73cadf314512460e7e04ff2a107c91e52516a
Reviewed-on: https://go-review.googlesource.com/c/build/+/529297
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>