internal/gomote: start gomote instance in swarming tasks
This change adds the ability to start a buildlet using a LUCI swarming
task. The swarming task will start the buildlet which connects to the
gomote server and authenticates against the rendezvous HTTP handler.
Once the handler authorizes the buildlet, it will initiate a reverse
dial connection with the gomote server. A buildlet client will be
created with the buildlet and will be stored in the session store for
use just as the previous gomote instances were.

This change also updates the version of Go used in the Dockerfile to
1.21..

Fixes golang/go#61843
Fixes golang/go#63781

Change-Id: I7eff163a200beb059265eb5c0f2c5ef7f4d6d911
Reviewed-on: https://go-review.googlesource.com/c/build/+/537456
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Auto-Submit: Carlos Amedee <carlos@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>