main.star: start adding a js-wasm builder
Start by adding it as a post-submit builder, only in the main Go repo,
and only at tip. Mostly so it's easier to see the generated code and
iterate. Once it works, it'll be turned up in the remaining places.

It turns out there's more complexity here because this is the first
case where the host OS differs from the target port, and we wish to
run the golangbuild binary on the host OS. For now, the target port
is being communicated by explicitly setting GOOS and GOARCH env vars.

For golang/go#61104.

Change-Id: I25aa946aeb16ef43b44322c25e133d19e68e6555
Reviewed-on: https://go-review.googlesource.com/c/build/+/530336
TryBot-Bypass: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Heschi Kreinick <heschi@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>