internal/task: add a task to update a test repo to follow release
For testing the module proxy, we always want to be able to download
modules that require the latest go toolchain. This task updates the
go directive of a go.mod of a test repository to include the latest
version that has been released.

This should be a post-release task for new major versions and minor
releases on the current major version.

This also updates the private sync task to correctly have
authentication configured.

Change-Id: Ic449e00db504301f08dcf6002bad7b67ca18c796
Reviewed-on: https://go-review.googlesource.com/c/build/+/533516
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>
Reviewed-by: Heschi Kreinick <heschi@google.com>