internal/workflow: change expansions to return values
Expansions were kind of fun for the smaller workflows, but had a major
flaw: there was no way to get data out of an expansion. Put another way,
once a workflow entered an expansion it basically couldn't leave.
That was okay as long as the larger workflows were static, but we're
about to switch to LUCI for testing, and at that point we can't compile
in the list of builders. I suppose we could read them in at startup,
but I'd like to do less of that kind of thing, not more.

Change them to allow returning a single Value. That might prove
limiting, but I'd really prefer not to have a combinatorial explosion
of expansion API just yet.

Unfortunately, I didn't see an easy way to allow this and still do
validation that there aren't orphaned tasks. In practice, even the most
trivial test should catch the problem, so meh. (We didn't used to have
stall detection, at which point it mattered much more.)

For golang/go#63147

Change-Id: I855523a72b87b0ed8089066f246e5869f28e6c89
Reviewed-on: https://go-review.googlesource.com/c/build/+/533860
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Heschi Kreinick <heschi@google.com>