dashboard: add OpenBSD 7.3 and 7.4 builders for 386, amd64
With the release of OpenBSD 7.4 we need to update our builders
configuration to support the newer versions, since the latest builder is
version 7.2.

I did not update the default builders for openbsd, openbsd-386 and
openbsd-amd64 to prevent current builders from failing, since we need
the GCE VMs ready.

Fixes [reserved]