main.star: adjust console names for the new builder groups page
The new builder groups page sorts by console "name", not by order of
definition like it previously did. Adjust the names so they're sorted
nicely again. This change adds an x- prefix to subrepo consoles, adds a
z- prefix to per-port consoles (to sort them last), and removes the -ci
suffix to get cleaner sorting within subrepo consoles.

Adjust the titles for better layout in the builder groups page as an
experiment. I have a theory that React is giving us shorter columns
because it can break up the text with whitespace in it. This change
removes that whitespace.

Change-Id: I109174c99a228972682d1b5e70c4e2951a4e0c2b
Reviewed-on: https://go-review.googlesource.com/c/build/+/537415
TryBot-Bypass: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>