main.star: simplify enabled
If a builder doesn't exist, it doesn't make sense for enabled to report
that it runs in pre- or postsubmit, as that makes using the function in
a loop harder.

Now that the loop for run mod policies is the only place that can set
exists to false, factor in the logic that cleared pre- and postsubmit
return values into that same loop to simplify the code.

Change-Id: Ie2b685ce11b16c73ad6a9ee7b582f24d54050928
Reviewed-on: https://go-review.googlesource.com/c/build/+/538896
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
TryBot-Bypass: Dmitri Shuralyov <dmitshur@golang.org>