internal/gomote, cmd/gomoteserver: use buildbucket for builder types
This change uses the build bucket API to retrieve builder data instead
of the configuration server API. It also changes the desired builder
type to the test only builders.

This also changes the script used to start the buildlet in a swarming
task. Python is already installed on all of the builders. Writing an
OS agnostic script to start the buildlet is preferred in Python.

For golang/go#63781
For golang/go#61773

Change-Id: I6137eac9f0525b6496b45b3ee46e46afaf34afa1
Reviewed-on: https://go-review.googlesource.com/c/build/+/540215
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Carlos Amedee <carlos@golang.org>