internal/task: fix Cloud Build some more
CL 534156 was written under the assumption that the Git checkout
went into a subdir of the workspace. I'm not sure if I somehow was wrong
about that, or if it's changed, but either way it's not the case now. So
do the Git work ourselves and make it work the way we want. Also fix the
output URLs, which were ending up looking like scratchdir/./go.mod. On a
real filesystem that'd be collapsed, but not on GCS.

This works in prod for x repo tagging and x509roots, which are the two
cases where we use Git checkouts I could test easily.

For golang/go#63147

Change-Id: I8a529b624d5c218a2940806fa6472149b4c82c9c
Reviewed-on: https://go-review.googlesource.com/c/build/+/539479
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Heschi Kreinick <heschi@google.com>