main.star: enable longtest in presubmit for protobuf
The protobuf repository (added in CL 540496) is special (i.e., it's
not one of the many golang.org/x repos that fits into a project type)
and configured to be tested only on Linux. Its owners find it helpful
for the longtest run mod to be enabled in presubmit by default.

After we finish taking care of the needs of special repositories and
see the patterns, we might want to refactor further.

For golang/go#63597.

Change-Id: I7edcefe8a49f1e89a497af390349412a32936462
Reviewed-on: https://go-review.googlesource.com/c/build/+/540278
Reviewed-by: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Bypass: Dmitri Shuralyov <dmitshur@golang.org>