main.star: configure host and target
The word "host" is getting confusing -- it now means both a machine type
and the platform that we're using to host the tests. For now I decided
to just live with it.

Per the CL description of https://crrev.com/c/5047151 we now have three
related but different concepts: the machine type, the host platform, and
the target platform.

Rename host_of to port_of, since that's more accurate to its intent, I
think? Then have host_of return the host platform for the build, and use
that to populate the host proto fields. host_of is where we map js-wasm
and android-amd64-emu to linux-amd64; dimensions_of still handles
running linux-386 builds on linux-amd64.

The dimensions for plan9 change intentionally -- now that we know the
BSDs can't run 32-bit binaries on 64-bit machines, we need the
386->amd64 swap to be more scoped.

Change-Id: I26bbd3fe3824aa4957b7068b3e8965e930e85a30
Reviewed-on: https://go-review.googlesource.com/c/build/+/544055
Auto-Submit: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Bypass: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>