main.star: add the staticlockranking builder
This change adds linux-amd64-staticlockranking builder. It's enabled in
postsubmit always, and in presubmit if any files in the runtime are
mutated. This builder is fairly important for avoiding deadlocks in the
runtime, and it's helpful to find out about this kind of breakage
sooner than postsubmit.

Change-Id: I815d4dd358b8026414008b53608c010da14f04aa
Reviewed-on: https://go-review.googlesource.com/c/build/+/546956
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
TryBot-Bypass: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>