main.star: formalize default host suffixes (OS versions)
This change builds on what CL 546955 started, which is a general
mechanism to apply a default suffix (OS version) to a host that doesn't
specify it in the builder type. Before we go that far, let's first
discuss what we mean by "host" in this circumstance.

We have 3 OS/arch combos we care about:
- Target platform: GOOS/GOARCH
- Host platform: GOHOSTOS/GOHOSTARCH
- Machine platform: the OS/arch of the machine we're actually running
  on.

Where we care about default host suffixes is the last one, and this is
what dimensions_of determines. host_of determines the second, and
split_builder_type determines the first, so they're the wrong places to
handle this.

With that out of the way, how do we handle it? It's simple: we require
all non-low-capacity builders to have a default host suffix. Normally
this information actually appears in the builder type and gets plumbed
down first through host_of, and then into dimension_of, but if the
information is missing then we fill it in. If we can't, we fail.

In a perfect world, we'd specify a precise OS version for every single
host. In practice, this isn't possible because we want low capacity
hosts to be defined as generally as possible, so that manual maintenance
performed on these hosts doesn't break them in CI. Oh well.

Note that because we only care about the machine OS/arch here, we don't
have to define default host suffixes for all platforms, only the ones we
have machines for that *aren't* low capacity.

Change-Id: I3c9e0f834d52231c13db7bc7fbddffa3e724c366
Reviewed-on: https://go-review.googlesource.com/c/build/+/547115
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Bypass: Michael Knyszek <mknyszek@google.com>