main.star: run longtest builders in presubmit if vendor is modified
This change enables longtest builders in presubmit if files related to
vendoring are modified. Having these enabled would've caught several
issues in vendored repositories and in the main Go repository in the
past.

While we're here, let's also ignore location filters for security
presubmit. Previously we applied them and this filtered down where the
staticlockranking builder would run, but if we apply them now then we'll
limit where longtest builders are run. However, we always want longtest
builders to run for security presubmit. Ignoring location filters
entirely is a reasonable decision here: we want to be as safe as
possible with security presubmit, so we want to test as many
configurations as possible. The extra builder runs are unlikely to cost
much and will help to ensure stability when landing security CLs
upstream.

For golang/go#42661.

Change-Id: Ic35216722ddca4c26cccbd96da045db5c100defb
Reviewed-on: https://go-review.googlesource.com/c/build/+/547596
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
TryBot-Bypass: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>