main.star: trust dry-runner dependencies
Treat unapproved dependencies owned by members of may-start-trybots as
trusted. This allows testing CLs in the middle of a stack owned by a
may-start-trybots user.

Note that CLs in the middle of a completely unprivileged user's CLs
still cannot be tested.

This option was added to LUCI in https://crrev.com/c/4879642 and
https://crrev.com/c/4879641.

Change-Id: I19f9b1c1a3059560063540c890fbdf1297d9dd52
Reviewed-on: https://go-review.googlesource.com/c/build/+/548477
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Bypass: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>