cmd/bootstrapswarm: remove -token-file-path flag
Reuse the swarming bot's existing mechanism for determining the path
to the token file. This should eliminate the possibility of it being
set for bootstrapswarm but not the swarming bot (or vice versa).

I also considered propagating the flag value to the swarming bot but
it seemed simpler to remove this flag for now.

Change-Id: I297ad9d8b931c9059c13fda08402b72693fa4488
Reviewed-on: https://go-review.googlesource.com/c/build/+/548955
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Than McIntosh <thanm@google.com>