env/windows-arm64/azure: fix problems with token generation setup
This patch revises the recipe for creating a scheduled task (via
powershell commands) to run the LUCI token generator. Turns out that
the default for scheduled tasks is to only run them when their
specified user is logged in, meaning that the recipe as written was
incorrect. The fix is to use the 'New-ScheduledTaskPrincipal' cmdlet
to boost the priority of the task and run it as a service account /
system user.

Updates golang/go#64587.

Change-Id: I281d8c5c11b0b41478524dfd456f4b1179c4d840
Reviewed-on: https://go-review.googlesource.com/c/build/+/549755
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>