cmd/gopherbot: document wiki page and where to see all tasks
Add a link to the gopherbot wiki page that documents some of the ways
that gopherbot can be interacted with.

We don't really have comprehensive documentation of all the gopherbot
tasks, and even if we did, it'd be hard to know if it's up to of date.
In practice, the source code is the best place to point people to for
now.

Add a note along those lines to the package comment.

Change-Id: Ifeb6e3875c1210417daa9c91d40567d0b928f580
Reviewed-on: https://go-review.googlesource.com/c/build/+/550377
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Sameer Ajmani <sameer@golang.org>
Reviewed-by: Sameer Ajmani <sameer@golang.org>