internal/task, relui: add posting to Mastodon
Very much a copy of the code for Twitter, except that it uses
a 3rd party package to handle posting.  Some of the common code
was moved into a new function.

Also includes an end-to-end test for by-hand use that will send a
Mastodon DM to a specified account.  This test requires access
to the cloud secrets and will skip if it lacks either those
secrets or the command line flag.

Fixes golang/go#65046.

Change-Id: Ie7606242766f36fecd5b15acc3c995c046098079
Reviewed-on: https://go-review.googlesource.com/c/build/+/554375
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>