main.star: set expiration_timeout on all builders
Currently the expiration timeout for a build is determined entirely by
buildbucket, but we've seen in practice that capacity constrained
builders in postsubmit can easily have builds sitting on the queue
waiting for a machine for longer than the buildbucket default (6h).

Set all builds targeting capacity constrained machines to have a 24h
expiration, and set every other build to have an expiration of 6h since
it seems to have worked fine so far.

We set the expiration timeout even on builders that don't need it for
consistency. A consequence of this is that we're no longer bound to
buildbucket's implementation default which can be a positive or a
negative depending on how you look at it. On the one hand we control the
expiration timeout now, which makes sense given that we know our fleet.
On the other hand, buildbucket may one day have an adaptive policy or
something that we miss out on. I think given that it hasn't happened
yet, let's err on the side of consistency; we can always change back.

Change-Id: I5daef457f434960158c93124d72adbd18a03d417
Reviewed-on: https://go-review.googlesource.com/c/build/+/554376
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
TryBot-Bypass: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>