main.star: reduce linux-amd64-ssacheck to a single shard
I noticed that the gotip-linux-amd64-ssacheck builder was creating 4
identical test shards that all run exactly the same command. This is
because `go tool dist test` doesn't understand how to shard test
packages for building. Also, it's pointless because just building the
test packages doesn't take very long at all.

Reduce the number of shards for this builder to 1 in the internal logic.
We could also make it an ALL_MODE builder, but that complicates the
logic without a huge benefit.

While we're here, let's simplify the logic for determining the test
shard count to more clearly enumerate all the possible cases.

Change-Id: I2744a899241d65570c6cc78e371f02b2232e0beb
Reviewed-on: https://go-review.googlesource.com/c/build/+/555095
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Bypass: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>