main.star: add linux-amd64-noopt builder type
This is a builder that exists in the old infrastructure but hasn't yet
been ported to LUCI. This is easy to do since it's just an environment
variable set throughout the build to turn off optimizations in builds.
This change adds a new run mod for noopt and the builder type.

Change-Id: I008527c9396808c63de8978e4023c3f3d387246c
Reviewed-on: https://go-review.googlesource.com/c/build/+/555096
TryBot-Bypass: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>