main.star: add linux-amd64-racecompile builder type
This is another builder that exists on the old infrastructure but not on
LUCI yet. The builder type rebuilds the compiler and linker in race mode
and then builds all of the standard library packages and commands.

This CL adds the new builder type and defines its corresponding run mod.
The run mod set a new property that was introduced to golangbuild as of
https://crrev.com/c/5182576.

Change-Id: Ice63253118e95713bee2207e66c71c3aba13c28a
Reviewed-on: https://go-review.googlesource.com/c/build/+/555097
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Bypass: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>