internal/task: update proxy task to check existing go version.
After the rc for the new go version has been tagged, there is a
tagged version > current major version, so the heuristics to choose
when to run the task do not apply. Instead, read the current go
version in the go.mod to find the old version and compare.

Change-Id: I2feefd9f4cb1e45e4e1d27f26b4416c392963955
Reviewed-on: https://go-review.googlesource.com/c/build/+/555675
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>