dashboard: disable x/build testing with Go 1.20
The x/build repo is permitted to be outside of the Go support policy
and has dropped support for Go 1.20 just now in CL 556355. We need to
backport that from LUCI to the old infrastructure while the migration
is underway and both dashboards are being supported.

Updating this in the old infra is a bit more involved, but this code
will be going away after the migration so I'm not trying to refactor.

For golang/go#64169.

Change-Id: Ic5079e38dd13b5bfb56d4d430efb5c3919918d47
Reviewed-on: https://go-review.googlesource.com/c/build/+/556455
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>