main.star: add -perf_vs_gotip and -perf_vs_parent builders
This change adds two run_mods that are used only in presubmit. This is
to begin initial testing of the performance builder, and these two
variants are fairly clearly going to be useful in the future.

This CL also moves up the definition of GO_BRANCHES in the file so that
it can be referenced from the RUN_MODS definitions.

Change-Id: Ifca845d833828baaea714ff25fb118d2a80b16c9
Reviewed-on: https://go-review.googlesource.com/c/build/+/556996
TryBot-Bypass: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>