main.star: wait for capacity on all capacity constrained builders
Capacity constrained builders tend to disappear occasionally for
manual maintainence or just because they're flaky. Without
wait_for_capacity set on the builder, when capacity reaches zero,
Swarming will fail the task as NO_RESOURCE.

I've observed this happening in practice for a few capacity constrained
platforms which results in infra failures. Fix this by always waiting
for capacity in capacity constrained settings.

Note: we also explicitly don't want to wait for capacity if we're *not*
capacity constrained. If our capacity ever hits zero with these
builders, then that likely indicates something is very wrong. We want to
be notified of that immediately, not after several hours.

Change-Id: I629fd115a5c7dcf0ebb5fb82c75915d4ceb86a77
Reviewed-on: https://go-review.googlesource.com/c/build/+/557856
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
TryBot-Bypass: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>