main.star: reorganize postsubmit consoles once again
This change was previous submitted as CL 557935 and subsequently
reverted because the change in console names meant history was lost. It
was also reverted because the console ordering is based on creation, not
console ID, so it didn't really make sense to submit it anyway. This is
the same CL but with those issues resolved.

This change reorganizes postsubmit consoles once again to make it easier
for the release rotation to identify issues.

Firstly, this CL drops the console headers. I think it was a nice idea,
but it's mostly visual noise.

Second, this CL adds the concept of known-issue builder types.

Third, this CL separates out known-issue builders into their own
consoles and puts links to the issues in the header.

Fourth, this CL unifies console creation in the make_console_gen
function and creates the notion of "tiers" for grouping consoles on the
console list page. Consoles are actually generated at the very end of
config generation so we can control the order in which they're created,
since console order in the builder groups list is ordered by the config
text.

Fifth, it moves the by-go-commit consoles into their own tier. On the
main console list page it generally just duplicates information but we
still want the links, and we want them to show up above known-issue
consoles.

And lastly, it updates the console titles once again to be more
human-friendly (for example, by replacing hyphens with whitespace).

Change-Id: I006904edb2c19d9f90f6c11f4a509028de482572
Reviewed-on: https://go-review.googlesource.com/c/build/+/558655
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Bypass: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>